Create useBuildings hook to access fetch buildings method and object type
